# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API docs for the golang project for learning.",
        "title": "Majestic gondola API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/tracks": {
            "get": {
                "description": "Get a list of all tracks in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "List tracks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TrackResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create multiple tracks at once from a JSON array",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Bulk create tracks",
                "parameters": [
                    {
                        "description": "List of tracks to create",
                        "name": "tracks",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CreateTrackRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrResponse"
                        }
                    }
                }
            }
        },
        "/tracks/populate/{count}": {
            "post": {
                "description": "Generate random tracks for development testing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dev"
                ],
                "summary": "Seed dummy data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of tracks to generate",
                        "name": "count",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Populate data",
                        "name": "track",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/PopulateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Invalid count",
                        "schema": {
                            "$ref": "#/definitions/ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrResponse"
                        }
                    }
                }
            }
        },
        "/tracks/{id}": {
            "get": {
                "description": "Retrieve a single track by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Get a track",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Track ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TrackResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "$ref": "#/definitions/ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Track not found",
                        "schema": {
                            "$ref": "#/definitions/ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the details of an existing track",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Update a track",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Track ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Track update data",
                        "name": "track",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateTrackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateTrackRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "artist_id": {
                    "type": "integer",
                    "example": 1
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Rock",
                        "Opera"
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "Bohemian Rhapsody"
                },
                "release_date": {
                    "type": "string",
                    "example": "1975-10-31 15:04:05"
                }
            }
        },
...
        "ErrResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "PopulateRequest": {
            "type": "object",
            "properties": {
                "artist_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "TrackResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Rock",
                        "Opera"
                    ]
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Bohemian Rhapsody"
                },
                "release_date": {
                    "type": "string",
                    "example": "1975-10-31 15:04:05"
                },
                "score": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "UpdateTrackRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "artist_id": {
                    "type": "integer",
                    "example": 1
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Rock",
                        "Opera"
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "Bohemian Rhapsody"
                },
                "release_date": {
                    "type": "string",
                    "example": "1975-10-31 15:04:05"
                }
            }
        },
...
    },
...
}